Fix user Dashboard page UI
