--- app/components/Buttons.js
@@ -1,15 +1,0 @@
-export const PrimaryButton = ({ label }) => {
-  return (
-    <button className="rounded-[30px] px-[20px] py-[10px] bg-[#1F4776] text-white text-sm">
-      {label}
-    </button>
-  );
-}
-
-export const SecondaryButton = ({ label }) => {
-  return (
-    <button className="rounded-[30px] px-[20px] py-[10px] bg-[#1F477610] text-[#1F4776] text-sm">
-      {label}
-    </button>
-  );
-}
