--- app/components/Data.js
@@ -1,31 +1,0 @@
-export const upcomingClasses = [
-  { name: "Vinyasa Yoga", date: "Monday, 14 February 2024" },
-  { name: "Pilates 1", date: "Wednesday, 16 February 2024" },
-];
-
-export const transactions = [
-  {
-    amount: -1,
-    date: "14/02/2024 01:00",
-    user: "Nara Smith",
-    description: "Booked ‘Vinyasa Yoga’ (16/02/2024 17:00)",
-  },
-  {
-    amount: -1,
-    date: "14/02/2024 01:00",
-    user: "Nara Smith",
-    description: "Booked ‘Hatha Yoga’ (18/02/2024 17:00)",
-  },
-  {
-    amount: 1,
-    date: "12/02/2024 01:00",
-    user: "Admin",
-    description: "Refund ‘ABT’ (12/02/2024 18:00)",
-  },
-  {
-    amount: 10,
-    date: "12/02/2024 01:00",
-    user: "Admin",
-    description: "Deposit",
-  },
-];
