--- app/components/SectionTitle.js
@@ -1,5 +1,0 @@
-export default function SectionTitle({ title }) {
-  return (
-    <p className="font-poppins font-bold text-2xl text-[#1F4776]">{title}</p>
-  );
-}
