+++ app/components/dashboard/ClassCard.js
@@ -1,17 +1,41 @@
-import { PrimaryButton } from "../Buttons";
+import PropTypes from "prop-types";
+import { format } from "date-fns";
+import ClassDetailsModal from "../classes/ClassDetailsModal";
+import { useDisclosure } from "@nextui-org/modal";
 
-export default function ClassCard({ name, date }) {
+export default function ClassCard({ booking }) {
+  const { isOpen, onOpen, onOpenChange } = useDisclosure();
+
   return (
-    <div className="flex flex-col gap-y-2.5 p-2.5 rounded-[20px] border-l-[4px] border-l-[#1F4776] bg-[#1F477610] text-[#393E46]">
-      <div className="flex flex-col">
-        <p className="font-bold text-base">{name}</p>
-        <p className="text-sm">{date}</p>
-        <p className="text-sm">UTown Gym Aerobics Studio</p>
+    <>
+      <div className="flex flex-col gap-y-2.5 p-2.5 rounded-[20px] border-l-[4px] border-l-a-navy bg-a-navy/10">
+        <div className="flex flex-col">
+          <p className="font-bold text-base">{booking.class.name}</p>
+          <p>{format(booking.class.date, "d/MM/y HH:mm")}</p>
+        </div>
+        <div className="flex justify-end">
+          {/* TODO: On clicking button, open modal similar to one in Classes with the details */}
+          <button
+            onClick={onOpen}
+            className="rounded-[30px] px-[20px] py-[10px] bg-a-navy text-white cursor-pointer"
+          >
+            View details
+          </button>
+        </div>
       </div>
-      <div className="flex justify-end">
-        {/* TODO: On clicking button, open modal similar to one in Classes with the details */}
-        <PrimaryButton label="View details" />
-      </div>
-    </div>
+      <ClassDetailsModal
+        selectedClass={booking.class}
+        selectedBooking={booking}
+        selected="booked"
+        userId={booking.userId}
+        isOpen={isOpen}
+        onOpen={onOpen}
+        onOpenChange={onOpenChange}
+      />
+    </>
   );
 }
+
+ClassCard.propTypes = {
+  booking: PropTypes.object,
+};
