--- app/components/logo.js
@@ -1,7 +1,0 @@
-export default function Logo() {
-  return (
-    <p className="font-display font-bold text-2xl text-[#1F4776] text-center">
-      NUS AEROBICS
-    </p>
-  );
-}
