+++ app/components/pages/AdminClassCreatePage.js
@@ -1,8 +1,8 @@
 import PropTypes from "prop-types";
 
 import { MdChevronLeft } from "react-icons/md";
-import AdminClassForm from "./AdminClassForm";
-import { PageTitle } from "../Titles";
+import ClassForm from "../classes/ClassForm";
+import { PageTitle } from "../utils/Titles";
 
 export default function AdminClassCreatePage({ closeCreate }) {
   return (
@@ -15,7 +15,7 @@ export default function AdminClassCreatePage({ closeCreate }) {
       </div>
       {/* TODO: On desktop, overflow should make entire white div longer rather than keep it at that height and scroll */}
       <div className="h-full w-full flex flex-col gap-y-5 p-5 overflow-scroll rounded-[20px] border border-a-black/10 bg-white">
-        <AdminClassForm
+        <ClassForm
           isCreate={true}
           selectedClass={{
             name: "",
