+++ app/dashboard/layout.js
@@ -1,6 +1,6 @@
-import NavBar from "../components/dashboard/navbar";
 import { redirect } from "next/navigation";
 import { getSession } from "../lib";
+import SideBar from "../components/dashboard/SideBar";
 
 export default async function Layout({ children }) {
   let user;
@@ -24,7 +24,7 @@ export default async function Layout({ children }) {
     // TODO: Make sure the overflow makes sense here too
     <div className="w-screen h-screen flex flex-col md:flex-row md:overflow-hidden">
       <div className="w-full flex-none md:w-64">
-        <NavBar user={user} />
+        <SideBar user={user} />
       </div>
       {/* TODO: Go through and ensure font sizing is standard */}
       <div className="w-full h-screen p-10 pt-20 bg-a-pink/80 gap-y-5 md:overflow-y-auto font-poppins text-a-black text-sm">
