+++ app/layout.js
@@ -3,9 +3,7 @@
 import Head from "next/head";
 import "../styles/globals.css";
 import Providers from "./providers";
-import { noto_serif_display, poppins } from "./components/fonts";
-
-const name = "NUS Aerobics";
+import { noto_serif_display, poppins } from "./components/utils/Font";
 
 export default function Layout({ children, home }) {
   return (
