+++ app/page.js
@@ -1,11 +1,3 @@
 export default function Home({}) {
   return <></>;
 }
-
-// import { redirect } from "next/navigation";
-// import { getSession } from "./lib";
-
-// export default async function Page() {
-//   const session = await getSession();
-//   return <p>{JSON.stringify(session, null, 2)}</p>;
-// }
